+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 * Settings > eholdings > Edit knowledge base credentials > Updating any field value generates an error message BUT the saves the update. (UIEH-1247)
 
+* Settings > eholdings > Edit knowledge base credentials > Updating any field value generates an error message BUT the saves the update. (UIEH-1247)
+* Include additional package content types in several responses. (UIEH-1240)
+
 ## [7.1.0] (https://github.com/folio-org/ui-eholdings/tree/v7.1.0) (2022-03-03)
 
 * Add tests for SettingsRoute component. (UIEH-1198)